amd64: fix the jmp_buf size

+++ include/arch-amd64/setjmp.h
@@ -26,20 +26,20 @@
 
 typedef struct	/* Do not change unless you also modify setjmp.S ! */
 {
-	unsigned int rbx;
-	unsigned int rcx;
-	unsigned int rdx;
-	unsigned int rsi;
-	unsigned int rdi;
-	unsigned int rbp;
-	unsigned int r8;
-	unsigned int r9;
-	unsigned int r10;
-	unsigned int r11;
-	unsigned int r12;
-	unsigned int r13;
-	unsigned int r14;
-	unsigned int r15;
-	unsigned int rsp;
-	unsigned int rip;
+	unsigned long rbx;
+	unsigned long rcx;
+	unsigned long rdx;
+	unsigned long rsi;
+	unsigned long rdi;
+	unsigned long rbp;
+	unsigned long r8;
+	unsigned long r9;
+	unsigned long r10;
+	unsigned long r11;
+	unsigned long r12;
+	unsigned long r13;
+	unsigned long r14;
+	unsigned long r15;
+	unsigned long rsp;
+	unsigned long rip;
 } jmp_buf[1];
